merge5
revisions 1

+++ public/js/sum2.js
@@ -1,7 +1,7 @@
 function sum2() {
       var uno = document.getElementById('rate').value;
       var dos = document.getElementById('days').value;
-      var ans = parseInt(dos)*(parseInt(uno));
+      var ans = parseFloat(dos)*(parseFloat(uno));
       if (!isNaN(ans)) {
          document.getElementById('basic').value = ans;
       }
